Make GmailFolder sync gmail labels
When synclabels config flag is set to "yes" for the gmail repo, offlineimap
fetches the message labels along with the messages, and embeds them into the
body under the header X-Keywords, as a comma separated list.

The configuration option labelsheader allows to change that header under which
labels are stored. X-Keywords is a useful choice as some mail programs may
recognize it.

It also adds an extra pass to savemessageto, that performs label synchronization
on existing messages from gmail to local, the same way it is done with flags.

The ignorelabels configuration seting contains is a list of comma separated
labels that will be left alone. They will not be added nor removed from any
message.